improved

+++ src/main/java/com/project/crm/dao/UserDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface  UserDao {
 
     public void addUser(User user);
-    public User getUserById(Long id);
+    public User getUserById(int id);
     public User getUserByUsername(String username);
     public List<User> getAllUsers();
     public boolean isExist(User user);
